+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 NOTES:
 
 BUG FIXES:
+* resource/hopsworksai_cluster: check if `ecr_registry_account_id` is not an empty string before setting it
 
 ENHANCEMENTS:
 * more unit tests